fix: rename context file
