+++ components/table/InternalTable.tsx
@@ -62,7 +62,7 @@ import type {
   TablePaginationPosition,
   TableRowSelection,
 } from './interface';
-import MeasureRowContext from './MeasureRowContext';
+import TableMeasureRowContext from './TableMeasureRowContext';
 import RcTable from './RcTable';
 import RcVirtualTable from './RcTable/VirtualTable';
 import useStyle from './style';
@@ -722,11 +722,11 @@ const InternalTable = <RecordType extends AnyObject = AnyObject>(
           transformColumns={transformColumns as any}
           getContainerWidth={getContainerWidth}
           measureRowRender={(measureRow) => (
-            <MeasureRowContext.Provider value={true}>
+            <TableMeasureRowContext.Provider value={true}>
               <ConfigProvider getPopupContainer={(node) => node as HTMLElement}>
                 {measureRow}
               </ConfigProvider>
-            </MeasureRowContext.Provider>
+            </TableMeasureRowContext.Provider>
           )}
         />
         {bottomPaginationNode}
